requirements.txt
flask==2.1.2
tensorflow==2.12.0
tensorflow-hub==0.13.0
Pillow==9.5.0
requests
